fix relative internal links on index pages
Index pages are always one level up comapred to other pages in a
directory. Relative links must reflect that.

So e.g. while a relative link on a regular page would be
`../another-dir/some-page`, on an index page, it would be
`./another-dir/some-page`.

For root index page, the current dir is `/<lang>/`, which means that a
rel link would be `./docs/another-dir/some-page`.

Which is all somewhat logical, but absolutely insane.